RouteWiseApi.Interests (phoenix_backend v0.1.0)

The Interests context - handles interest categories and user interests. This is a convenience module that delegates to the Trips context.

Functions

create_interest_category(attrs \\ %{})

Creates an interest category.

create_user_interest(attrs \\ %{})

Creates a user interest.

delete_interest_category(interest_category)

Deletes an interest category.

delete_user_interest(user_interest)

Deletes a user interest.

get_interest_category!(id)

Gets a single interest category.

get_interest_category_by_name(name)

Gets an interest category by name.

get_user_interests(user_id)

Returns the list of user interests for a specific user.

list_interest_categories()

Returns the list of interest categories.

update_interest_category(interest_category, attrs)

Updates an interest category.

update_user_interest(user_interest, attrs)

Updates a user interest.