RouteWiseApi.Places (phoenix_backend v0.1.0)

The Places context for managing place data and Google Places API integration.

Functions

change_place(place, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking place changes.

Examples

iex> change_place(place)
%Ecto.Changeset{data: %Place{}}

cleanup_old_cache(list)

Cleans up old cached place data.

Examples

iex> cleanup_old_cache(hours: 48)
{5, nil}

create_place(attrs \\ %{})

Creates a place.

Examples

iex> create_place(%{field: value})
{:ok, %Place{}}

iex> create_place(%{field: bad_value})
{:error, %Ecto.Changeset{}}

create_place_from_google(google_data)

Creates a place from Google Places API data.

Examples

iex> create_place_from_google(google_data)
{:ok, %Place{}}

iex> create_place_from_google(invalid_data)
{:error, %Ecto.Changeset{}}

delete_place(place)

Deletes a place.

Examples

iex> delete_place(place)
{:ok, %Place{}}

iex> delete_place(place)
{:error, %Ecto.Changeset{}}

get_place(id)

Gets a single place by ID.

Returns nil if the Place does not exist.

Examples

iex> get_place(123)
%Place{}

iex> get_place("invalid")
nil

get_place!(id)

Gets a single place by ID.

Raises Ecto.NoResultsError if the Place does not exist.

Examples

iex> get_place!(123)
%Place{}

iex> get_place!("invalid")
** (Ecto.NoResultsError)

get_place_by_google_id(google_place_id)

Gets a place by Google Place ID.

Examples

iex> get_place_by_google_id("ChIJN1t_tDeuEmsRUsoyG83frY4")
%Place{}

iex> get_place_by_google_id("nonexistent")
nil

get_places_by_type(location, place_type, radius \\ 5000)

Gets cached places by type within a location.

Examples

iex> get_places_by_type(%{lat: 37.7749, lng: -122.4194}, "restaurant")
[%Place{}, ...]

list_places()

Returns the list of places.

Examples

iex> list_places()
[%Place{}, ...]

search_places_near(location, query \\ nil, radius \\ 5000)

Searches for places by location and query.

Examples

iex> search_places_near(%{lat: 37.7749, lng: -122.4194}, "restaurant")
[%Place{}, ...]

update_place(place, attrs)

Updates a place.

Examples

iex> update_place(place, %{field: new_value})
{:ok, %Place{}}

iex> update_place(place, %{field: bad_value})
{:error, %Ecto.Changeset{}}

upsert_place_from_google(google_data)

Updates or creates a place from Google Places API data.

Examples

iex> upsert_place_from_google(google_data)
{:ok, %Place{}}